requirements.txt
pandas>=1.3
numpy
pyarrow
pandera
pyyaml
pytest
dvc